+++ CHANGELOG.md
@@ -10,6 +10,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 - FIXED: minor cfdb::access issue leading to catalog build failure in some configurations
 - FIXED: to more reliably detect active version of PostgreSQL for backup purposes
 - FIXED: to bundle pg_backup_ctl with custom changes for PostgreSQL v10+
+- FIXED: PostgreSQL to always use the original config
 - CHANGED: updated PostgreSQL latest version to v11
 - CHANGED: to use cfbackup module